refactor: Move modal methods from DialogUtil to ModalUtil class
