+++ lib/presentation/pages/generate/widgets/image_selector.dart
@@ -1,5 +1,6 @@
 import 'dart:io';
 
+import 'package:cooki/core/utils/modal_util.dart';
 import 'package:easy_image_viewer/easy_image_viewer.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
@@ -8,7 +9,6 @@ import 'package:flutter_svg/svg.dart' show SvgPicture;
 import 'package:image_picker/image_picker.dart';
 
 import '../../../../app/constants/app_colors.dart';
-import '../../../../core/utils/dialogue_util.dart';
 import '../../../../core/utils/logger.dart';
 import '../generate_recipe_view_model.dart';
 
@@ -21,7 +21,7 @@ class ImageSelector extends ConsumerWidget {
     BuildContext context,
     GenerateRecipeViewModel vm,
   ) {
-    DialogueUtil.showImagePickerModal(
+    ModalUtil.showImagePickerModal(
       context,
       onCamera: () => _pickImage(vm, ImageSource.camera),
       onGallery: () => _pickImage(vm, ImageSource.gallery),
